+++ Magellan/Classes/Modules/Review/Comments/CommentViewModel.swift
@@ -1,0 +1,28 @@
+//
+/**
+* Copyright Soramitsu Co., Ltd. All Rights Reserved.
+* SPDX-License-Identifier: GPL-3.0
+*/
+
+
+import Foundation
+
+struct CommentViewModel<Cell: CommentTableViewCell>: CommentViewModelProtocol {
+    
+    let style: MagellanStyleProtocol
+    let fullName: String
+    let rate: Double
+    let date: Date
+    let text: String
+    var avatarURL: String?
+    var shortTitle: String? { title.shortUppercased }
+        
+    var cellType: UITableViewCell.Type { Cell.self }
+    
+    func bind(to cell: UITableViewCell) {
+        (cell as? Cell).map {
+            $0.bind(viewModel: self)
+            Cell.Default(style: style).apply(to: $0)
+        }
+    }
+}
