+++ Magellan/Classes/Modules/Review/Extensions/UIView+RoundShape.swift
@@ -1,0 +1,20 @@
+//
+/**
+* Copyright Soramitsu Co., Ltd. All Rights Reserved.
+* SPDX-License-Identifier: GPL-3.0
+*/
+
+
+import Foundation
+
+extension UIView {
+    
+    func makeRound() {
+        let mask = CAShapeLayer()
+        let path = UIBezierPath(ovalIn: bounds)
+        mask.path = path.cgPath
+        mask.fillRule = .evenOdd
+        layer.mask = mask
+    }
+    
+}
